Make UidReverseComparator a top-level class
