+++ k9mail/src/main/java/com/fsck/k9/controller/MessagingController.java
@@ -7,7 +7,6 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.Date;
 import java.util.EnumSet;
 import java.util.HashMap;
@@ -133,34 +132,6 @@ public class MessagingController implements Runnable {
     private static final String PENDING_COMMAND_APPEND = "com.fsck.k9.MessagingController.append";
     private static final String PENDING_COMMAND_MARK_ALL_AS_READ = "com.fsck.k9.MessagingController.markAllAsRead";
     private static final String PENDING_COMMAND_EXPUNGE = "com.fsck.k9.MessagingController.expunge";
-
-    public static class UidReverseComparator implements Comparator<Message> {
-        @Override
-        public int compare(Message messageLeft, Message messageRight) {
-            Integer uidLeft, uidRight;
-            try {
-                uidLeft = Integer.parseInt(messageLeft.getUid());
-            } catch (NullPointerException | NumberFormatException e) {
-                uidLeft = null;
-            }
-            try {
-                uidRight = Integer.parseInt(messageRight.getUid());
-            } catch (NullPointerException | NumberFormatException e) {
-                uidRight = null;
-            }
-
-            if (uidLeft == null && uidRight == null) {
-                return 0;
-            } else if (uidLeft == null) {
-                return 1;
-            } else if (uidRight == null) {
-                return -1;
-            }
-
-            // reverse order
-            return uidRight.compareTo(uidLeft);
-        }
-    }
 
     /**
      * Maximum number of unsynced messages to store at once
